Remove manifest timestamps because the content based hash strategy renders them unnecessary.
