+++ src/app/api/manifest/__tests__/handler.test.ts
@@ -1,17 +1,7 @@
-import { beforeAll, describe, expect, it } from "vitest";
-import { computeFirstChoiceBreakdown } from "@/packages/contracts/slices/first_choice_breakdown/compute";
-import { ingestCvr } from "@/packages/contracts/slices/ingest_cvr/compute";
-import { computeStvRounds } from "@/packages/contracts/slices/stv_rounds/compute";
+import { describe, expect, it } from "vitest";
 import { handleManifestRequest } from "../handler";
 
 describe("handleManifestRequest", () => {
-  beforeAll(async () => {
-    // Generate test data
-    await ingestCvr();
-    await computeFirstChoiceBreakdown();
-    await computeStvRounds();
-  });
-
   it("should return success with manifest data", async () => {
     const result = await handleManifestRequest("test");
 
@@ -41,11 +31,9 @@ describe("handleManifestRequest", () => {
     expect(result.data).toHaveProperty("version");
     expect(result.data).toHaveProperty("env");
     expect(result.data).toHaveProperty("elections");
-    expect(result.data).toHaveProperty("generated_at");
     expect(result.data).toHaveProperty("inputs");
     expect(typeof result.data?.version).toBe("number");
     expect(typeof result.data?.env).toBe("string");
-    expect(typeof result.data?.generated_at).toBe("string");
   });
 
   it("should include contest structure with required properties", async () => {
@@ -136,23 +124,6 @@ describe("handleManifestRequest", () => {
     const election = result.data?.elections[0];
     expect(election?.election_id).toBe("portland-20241105-gen");
     expect(election?.contests[0]?.contest_id).toBe("d2-3seat");
-  });
-
-  it("should have consistent timestamps", async () => {
-    const result = await handleManifestRequest("test");
-
-    expect(result.success).toBe(true);
-
-    // Check that generated_at is a valid ISO string
-    const generatedAt = result.data?.generated_at;
-    expect(generatedAt).toBeDefined();
-    expect(() => new Date(generatedAt!)).not.toThrow();
-
-    // Check that the timestamp is recent (within last hour)
-    const timestamp = new Date(generatedAt!);
-    const now = new Date();
-    const oneHourAgo = new Date(now.getTime() - 60 * 60 * 1000);
-    expect(timestamp.getTime()).toBeGreaterThanOrEqual(oneHourAgo.getTime());
   });
 
   it("should handle consistent data structure across multiple calls", async () => {
