+++ src/lib/manifest/__tests__/contest-resolver.test.ts
@@ -9,7 +9,6 @@ describe("ContestResolver", () => {
   const testManifest: Manifest = {
     env: "test",
     version: 2,
-    generated_at: "2024-01-01T00:00:00.000Z",
     inputs: {},
     elections: [
       {
