+++ src/lib/manifest/__tests__/loaders.test.ts
@@ -152,7 +152,6 @@ describe("loaders", () => {
     testManifest = {
       env: "test",
       version: 2,
-      generated_at: "2024-01-01T00:00:00.000Z",
       inputs: {},
       elections: [
         {
